Basic server implementation (#2)
* supports get/fetch/update/exec
* binary protocol only

+++ src/ThingSet.Common/Protocols/Binary/CborDeserialiser.cs
@@ -21,7 +21,7 @@ public static class CborDeserialiser
             case CborReaderState.UnsignedInteger:
                 return ReadInteger(reader, state);
             case CborReaderState.HalfPrecisionFloat:
-                return reader.ReadHalf();
+                return (float)reader.ReadHalf();
             case CborReaderState.SinglePrecisionFloat:
                 return reader.ReadSingle();
             case CborReaderState.DoublePrecisionFloat:
